+++ compiler/src/printExportInfo.mli
@@ -1,0 +1,17 @@
+open Prog
+
+val pp_export_info :
+  ?json:bool ->
+  Format.formatter ->
+  ( unit,
+    ( 'reg,
+      'regx,
+      'xreg,
+      'rflag,
+      'cond,
+      'asm_op,
+      'extra_op )
+    Arch_extra.extended_op )
+  prog ->
+  ('reg, 'regx, 'xreg, 'rflag, 'cond, 'asm_op) Arch_decl.asm_prog ->
+  unit