+++ CHANGELOG.md
@@ -82,6 +82,12 @@
   ([PR #1083](https://github.com/jasmin-lang/jasmin/pull/1083);
   fixes [#1079](https://github.com/jasmin-lang/jasmin/issues/1079)).
 
+- Added two new options `-print-export-info` and `-print-export-info-json` to
+  display interesting information about exported functions, in readable and
+  json formats. The readable format only filters out important information
+  (for the moment, if the argument alignment is other than 8).
+  ([PR #1123](https://github.com/jasmin-lang/jasmin/pull/1123)).
+
 ## Bug fixes
 
 - Improve heuristic for relational variables in safety analysis